Initial source priming
The OpenPOWER Foundation master template document is based loosely
on the OpenStack Manuals project in GitHub at
https://github.com/openstack/openstack-manuals

Many of the examples in the documentation were created by another IBM
employee, Jeff Brown (jeffdb@us.ibm.com).  Reviews of the content
have been handled by the OpenPOWER Foundation Technical Steering
Committee and the System Software Work Group.

While the format of the document is in good shape, the technical details
remain to be verified and the Docbook source reviewed before publishing
version 1.0.  However, this is a good starting point.

Signed-off-by: Jeff Scheel <scheel@us.ibm.com>