Restructure Project to use Docs-Master Project
This project has now been built to longer contain common files nor the
master POM file and now instead relies on the ../Docs-Master/ structure
when building.  As such, the doc/common directory was removed and the
remaining files in doc/ were promoted to main level.  Then, the top
POM file was converted to a "minimal" POM referencing the master-pom
in ../Docs-Master/.  Finally, the template/ directory required a
slight update to the POM file (new name for parent minimal POM,
"workgroup-pom") and updated locations (../../Docs-Master/common) for
the common preface and appendix files in the book file (bk_main.xml).

For completeness, the README file was updated to ensure users of the
project understand it no longer is self-contained and must rely on the
Docs-Master project.

Signed-off-by: Jeff Scheel <scheel@us.ibm.com>